### 2017-10-29 #444 @whedon
[PRE REVIEW]: PolaronMobility.jl: Implementation of the Feynman variational polaron model
**Submitting author:** @jarvist (<a href="http://orcid.org/0000-0003-1938-4430">Jarvist Moore Frost</a>)
**Repository:** <a href="https://github.com/jarvist/PolaronMobility.jl" target ="_blank">https://github.com/jarvist/PolaronMobility.jl</a>
**Version:** ArXiv-v4
