### 2017-10-29 #444 @jarvist
From your list of available reviewers:
@jochym and @khinsen both have more than sufficient electronic structure background to understand where these codes are coming from, but do not list any Julia experience. The codes aren't that complex, or use that specific language features (metaprogramming etc.), but it might be nice to have some expertise & feedback from scientific Julia users, in how to (re)structure the package.

@bjmorgan also has the scientific background -- but I know him & we're both employees at the University of Bath!

Of the reviewers who indicate Julia familiarity, perhaps @lheagy is the closest in scientific overlap.
