### 2017-10-30 fb20668
Some very basic tests of the Feynman Alpha parameter; c.f. literature
