### 2017-10-30 5fd1d2a
More tests; separate tests into their own files.
