+++ test/HellwarthMultipleBranches.jl
@@ -1,0 +1,141 @@
+# HellwarthEffectiveFrequency.jl 
+#   - use Hellwarth et al. 1999 PRB method to reduce multiple phonon modes to a single effective frequency
+
+push!(LOAD_PATH,"../src/") # load module from local directory
+using PolaronMobility 
+
+# ((freq THz)) ((IR Activity / e^2 amu^-1))
+# These data from MAPbI3-Cubic_PeakTable.csv
+# https://github.com/WMD-group/Phonons/tree/master/2015_MAPbI3/SimulatedSpectra
+# Data published in Brivio2015 (PRB)
+# https://doi.org/10.1103/PhysRevB.92.144308
+MAPI= [
+96.20813558773261 0.4996300522819191
+93.13630357703363 1.7139631746083817
+92.87834578121567 0.60108592692181
+92.4847918585963 0.0058228799414729
+92.26701437594754 0.100590086574602
+89.43972834606603 0.006278895133832249
+46.89209141511332 0.2460894564364346
+46.420949316788 0.14174282581124137
+44.0380222871706 0.1987196948553428
+42.89702947649343 0.011159939465770681
+42.67180170168193 0.02557751102757614
+41.46971205834201 0.012555230726601503
+37.08982543385215 0.00107488277468418
+36.53555265689563 0.02126940080871224
+30.20608114002676 0.009019481779712388
+27.374810898415028 0.03994453721421388
+26.363055017011728 0.05011922682554448
+9.522966890022039 0.00075631870522737
+4.016471586720514 0.08168931020200264
+3.887605410774121 0.006311654262282101
+3.5313112232401513 0.05353548710183397
+2.755392921480459 0.021303020776321225
+2.4380741812443247 0.23162784335484837
+2.2490917637719408 0.2622203718355982
+2.079632190634424 0.23382298607799906
+2.0336707697261187 0.0623239656843172
+1.5673011873879714 0.0367465760261409
+1.0188379384951798 0.0126328938653956
+1.0022960504442775 0.006817361620021601
+0.9970130778462072 0.0103757951973341
+0.9201781906386209 0.01095811116040592
+0.800604081794174 0.0016830270365341532
+0.5738689505255512 0.00646428491253749
+#0.022939578929507105 8.355742795827834e-05   # Acoustic modes!
+#0.04882611767873102 8.309858592685e-06
+#0.07575149723846182 2.778248540373041e-05
+]
+
+MAPbBr3 = [
+#v / THz IR Activity / e^2 amu^-1
+96.25494581101505 0.4856105419754306
+93.36827101597564 1.5237045178260684
+93.19828072025739 0.54951133126244
+92.59570097621418 0.0017497942636829
+92.36075758752216 0.054483693922451705
+89.49998024367731 0.003950754031148851
+47.175408954380885 0.21156964115237242
+46.57576693654983 0.11627478349663156
+44.3361503231313 0.19163912283684897
+42.967352451897554 0.011987727827091217
+42.71066583604171 0.027797225985699996
+41.490751582679316 0.01234463397562453
+37.22033177431264 0.001649578352399309
+36.57010484641483 0.019161684306162618
+30.438999700507342 0.007365167772350685
+27.625465570185764 0.035161502048866464
+26.458785163031635 0.038174133532805965
+9.947074760143552 0.0012355225072173002
+4.493935108724231 0.07714630445769342
+4.215763207131703 0.01724042408702847
+4.021795301113039 0.06167913225780508
+3.2393471899404047 0.040335730459227545
+2.914057478007975 0.27361583921530497
+2.6223962950449407 0.23493646364819373
[…]
