### 2017-10-30 8d31bcc
BUGFIX: Don't generate bogus .dat files automatically.
