### 2017-10-31 b78d9e5
Moved Susceptibility to own file; incorporated OedipusRex.jl:
https://github.com/jarvist/OedipusRex
