### 2017-10-31 8cfe0ce
Move PlotPolaron to own module/file; freeing Plots dep.
