+++ test/runtests.jl
@@ -10,5 +10,5 @@ include("FeynmanAlpha.jl")
 include("HellwarthMultipleBranches.jl")
 include("FrostPolaronMobility2017.jl")
 
-println("That's me! If I finished without interupting, let's assume the tests were A.OK."
+println("That's me! If I finished without interupting, let's assume the tests were A.OK.")
 
