+++ src/PolaronMobility.jl
@@ -1,7 +1,7 @@
 module PolaronMobility
 
 include("FeynmanKadanoffOsakaHellwarth.jl")
-include("susceptibility.jl") # ImX calculation
+include("Susceptibility.jl") # ImX calculation
 include("OedipusRex.jl") # Optical Absorption
 
 end # module
