### 2017-10-31 ceaa2c1
Bugs! Not sufficient testing before prev. commit...
