### 2017-11-03 8e87e40
Clean up polaronmobility call signature; propagate changes.
Now that savepolaron + plotpolaron are (sensibly) separate functions,
polaronmobility has no need to know fileprefix.

Also added docstring for polaron mobility.
