### 2017-11-06 d90f2c1
Extra comments, code cleanup.
