### 2017-11-09 983153e
Correct signature for documentation; latest docstrings.
