### 2017-11-10 cd10e3d
Basic running of the susceptibility (freq dep mob code).
