### 2017-12-09 #444 @arfon
@lheagy - would you be willing to edit this submission for JOSS?
