### 2017-12-11 #444 @lheagy
@arfon, @jarvist: this is outside of my background, but I am happy to help.
