### 2017-12-11 #444 @lheagy
@jarvist: I have looked through the paper and briefly through your software. I have a couple suggestions before we send it out for review.

## Paper
- [ ] the paper is quite technical, do you have a couple examples or applications you could discuss briefly in order to reach a wider audience?
- [ ] you can remove the "Spiel from original website below"
- [ ] citations should be in square brackets (eg.
```
This package is an implementation in the Julia programming language of solving
the Feynman [@Feynman1955]
```

## Software
- [ ] it looks like the testing and code coverage are not currently set up (the badge links at the top of the Readme are all unknown: https://github.com/jarvist/PolaronMobility.jl/blob/master/README.md)
