### 2017-12-12 #444 @jarvist
Thanks for the feedback! I'll try and get those edits done today or tomorrow, before it's sent out.
