### 2017-12-20 53a688a
Update README; following JOSS guidance:
https://github.com/openjournals/joss-reviews/issues/444#issuecomment-350792479
