+++ README.md
@@ -13,14 +13,25 @@ More extensive documentation is available
 which is perhaps easiest to read and understand alongside the pre-print:
 [ArXiv:1704.05404](https://arxiv.org/abs/1704.05404) .
 
-The scientific inputs are the dielectric constants, a characteristic phonon
-frequency, and the bare-electron band effective-mass.
+The required inputs are the dielectric constants (ϵ-static and ϵ-optic)
+, a characteristic phonon frequency (ω), and the bare-electron band
+effective-mass (me).  These values can be relatively easily calculated in the
+ab-initio electronic structure package of your choosing, or measured directly. 
 
-From this you can solve a temperature-dependent polaron model (variationally
-optimising temperature-dependent free-energies for the coupled electron-phonon
-system), and from this optimised parameters calculate derivative quantities
-such as the polaron mobility, and polaron features (effective masses, energy
-loss rates, oscillation etc.)
+From these four values, the code solves a temperature-dependent polaron model. 
+This is done by variationally optimising the temperature-dependent free-energies for the coupled
+electron-phonon system. 
+These optimised parameters describe the polaron with the infinite quantum field
+of lattice vibrations 'integrated through', and replaced with a phonon-drag
+term. 
+From this the polaron features such as effective-mass, size of the
+wavefunction, frequency of energy oscillation etc. can be calculated. 
+
+This polaron state can then be used as an input to further models for polaron
+mobility. 
+
+The codes are designed to produce a set of temperature-dependent mobilities and
+other data, for direct incorporation into a scientific publication. 
 
 May your phonons drag in a manner truly sublime.
 
@@ -37,6 +48,8 @@ julia> Pkg.clone("git://github.com/Jarvist/PolaronMobility.jl.git")
 ```
 
 ## Research outputs
+
+As an example, here are some indicative figures generated by the code. 
 
 Polaron mobilities, three different ways
 ![Polaron mobilities, three different ways](mobility-calculated.png)
