### 2017-12-20 0c41c72
Additions / corrections to the documentation.
