### 2017-12-20 aeb36fa
Minor clean + rewrite.
