### 2017-12-20 #444 @jarvist
Thanks for the timely reminder :^)

Had to have a read about before getting CI working.

I hope I've made the text a little bit more approachable.
