### 2017-12-29 #444 @lheagy
Thanks @jarvist: I had a look and the paper and it doesn't look like the most recent changes have made it to your master branch (the last update was October). Is it on a branch somewhere? Thanks!

![image](https://user-images.githubusercontent.com/6361812/34442023-9cf1336e-ec7c-11e7-865e-0c9066e02f96.png)
