### 2018-01-05 #444 @jarvist
My apologies! I had managed to confuse myself + ended up tidying the README instead of the JOOS paper.
I've applied edits as suggested & tried to write a more lay-person introduction.

https://github.com/jarvist/PolaronMobility.jl/commit/215cc55d63ad233e79f7c8cba79835ff0fad754d
