### 2018-01-07 #444 @lheagy
Thanks @jarvist, the text looks good. There is an issue with the bibliography when trying to generate the pdf - do you need the citation on line 9 at the moment? It doesn't look like you reference it, so we could potentially remove it for now?

How are things going with getting the testing up and running?
