### 2018-01-09 6206ddf
Remove blank bib entry; issue with JOSS rendering.
