+++ JOSS-paper/paper.bib
@@ -5,9 +5,6 @@
     author = {Jarvist Moore Frost}
 } 
 
-# For Zenodo / JOSS DOI to the software package (GitHub)
-@misc{PolaronMobility.jl} 
-
 % Polaron papers 
 
 # Feynman model polaron fundamentals 
