### 2018-01-09 #444 @jarvist
Ok, I've removed the blank reference.

Travis CI is setup, but the builds sometimes fail due to reasons outside my control - often that Travis fails to get a fresh copy of Julia!

I'll see if I can restrict the architectures it tries on;
https://travis-ci.org/jarvist/PolaronMobility.jl/jobs/325561158
