### 2018-01-11 #444 @lheagy
As this submission is outside my domain, I would like to find 2 reviewers.

@tacaswell, @khinsen, @betatim, @jochym - would any of you be willing to review this submission for JOSS? You can read more about the process here http://joss.theoj.org/about#reviewer_guidelines, and feel free to ping with questions!
