### 2018-01-21 #444 @lheagy
👋 Hi @tacaswell, @khinsen, @betatim, @jochym - would any of you be willing to review this submission for JOSS?
