### 2018-01-29 #444 @lheagy
Hi @dean0x7d, @ajgpitch, would you be willing to review this submission on PolaronMobility.jl  for [JOSS](Joss.theoj.org/about)? JOSS is a 'developer friendly' journal. That is, we focus on the software as a scientific contribution. If the submitting authors have followed best practices (have documentation, tests, continuous integration, and a license) then their review should be straight-forward. There are more details on reviewing with JOSS here: http://joss.theoj.org/about#reviewer_guidelines

With your background in solid state physics, I would greatly appreciate your input!
