### 2018-02-01 #566 @whedon
[REVIEW]: PolaronMobility.jl: Implementation of the Feynman variational polaron model
**Submitting author:** @jarvist (<a href="http://orcid.org/0000-0003-1938-4430">Jarvist Moore Frost</a>)
**Repository:** <a href="https://github.com/jarvist/PolaronMobility.jl" target ="_blank">https://github.com/jarvist/PolaronMobility.jl</a>
**Version:** ArXiv-v4

- [x] **Version:** Does the release version given match the GitHub release (ArXiv-v4)?
