### 2018-02-11 jarvist/PolaronMobility.jl#1 @lheagy
JOSS Review: Documentation
I think there are a number of steps that can be taken in the documentation to help users get up and running more easily. This is connected to the JOSS review at openjournals/joss-reviews#566

- [x] First off, the documentation is built as a single, large page, if this could be split into multiple pages, that would help ease navigation
- [x] Installation instructions: these are in the README, but should also be included in the documentation
- [x] Example usage: in addition to the plots shown in the readme, it would be helpful to include examples which show both the code usage and the resulting plot

@yxqd may also have some suggestions here.
