### 2018-02-11 #566 @lheagy
Hi @jarvist, I have added an issue regarding documentation based on the guidelines above (jarvist/PolaronMobility.jl#1) Please have a look and let me know if there is anything I can clarify. @yxqd may have some comments to add here as well.
