### 2018-02-12 #566 @yxqd
Thanks @lheagy. I agree with your comments. Here are some additional points:

Hi @jarvist, your software looks very interesting. There are a few things to fix but it is mostly about documentation. As a first time user, after I follow your installation instruction on your README, it is unclear to me what is the next step, and how to learn to use your software. It would be really helpful to have a detailed example (tutorial), including for example brief explanation of how the parameters (ϵ-static and ϵ-optic, for example) were obtained. Also please read the first post in this issue, where a checklist of requirements is presented. For example, "community guidelines" is missing from your software documentation.

It is great that you are using travis-CI to automatically test the software! And it would be nice to increase the coverage to a healthier state.
