### 2018-02-12 jarvist/PolaronMobility.jl#1 @yxqd
@jarvist: I agree with @lheagy. For the third point above on "Example usage", a tutorial will be really helpful. And I agree that you should expand the documentation site to add sections on installation and usage, and improve the API section.
