### 2018-02-20 jarvist/PolaronMobility.jl#1 @jarvist
These all sound sensible points. I'll see what I can do. Thanks.
