### 2018-02-20 94b75db
JOSS-review motivated improvements to documentation; first draft.
