+++ docs/src/scientific_discussion.md
@@ -1,0 +1,237 @@
+## Overview 
+
+These codes solve Osaka's [Osaka1961] finite-temperature parameters for the Feynman
+polaron model [Feynman1955]. 
+We copy the form used in [Hellwarth1999].
+
+For each temperature, the total free energy of the Feynman coupled
+phonon-electron system is minimised by optimising coefficients which control
+the spring-coupling coefficient and effective-mass for the phonon cloud. 
+You have integrated through the infinite quantum field of the harmonic
+oscillators which make up the lattice, to collapse the problem abck to
+a quasi-particle. 
+
+The codes then calculate the polaron mobility, with both the original low-temperature
+FHIP approximation [Feynman1962]; Kadanoff's [Kadanoff1964] Boltzmann equation
+motivated phonon-emission correction to the FHIP; and Hellwarth et al.'s
+[Hellwarth1999] method. 
+This last method (Hellwarth) is probably the most accurate. 
+It uses a more general (non low-temperature-approximation) result in
+[Feynman1962], then directly calculates the contour integral for the polaron
+self-energy with numerical integration.
+
+Underlying all this is the simplified Frohlich Hamiltonian [Frohlich1952] for
+a single electron interacting with a phonon cloud of non-interacting (harmonic)
+phonons.
+The electron-phonon interaction for a polar system is treated at the simple
+level of being the dielectric response. 
+These are the infrared-active modes present at the Gamma point in the Brillouin
+zone. 
+Along with an effective mode frequency, the dielectric constants are used to
+calculate the dimensionless 'α' parameter describing the electron-phonon
+coupling.  
+(In a simple covalent semiconductor system, the only dielectric active mode is
+the linear-optical mode.)
+
+The Feynman model offers a direct solution of this most simple quantum field
+problem. 
+The the infinite phonon (quantum) field is 'integrated out' by path
+integration. 
+The soluble system is one in which you have an electron interacting by
+a (harmonic) spring constant with a mass representing the phonon drag. 
+The variational method allows you to find a set of parameters for this
+simplified system which produces the smallest free-energy. 
+
+This is now a (renormalised) single particle system, a quasi-particle.
+
+These Julia codes use Hellwarth's [Hellwarth1999] presentation of Osaka's variational
+free-energies for the Feynman model.
+We optimise the `v` and `w` parameters for these finite-temperature free energies.
+These can be alternatively restated the mass 'M' and spring-constant 'k' of the
+coupled phonon-electron Feynman model.
+
+Here we apply these methods to the case of hybrid halide perovskites.
+The method provides the temperature dependent polaron-mobility without any free parameters.
+No arbitrary relaxation time is needed or used. The scattering processes are
+treated directly, by including an effective electron-phonon coupling in the
+specification of the Frohlich 'alpha' parameter, and then all other features
+come from solving the model.
+The original Feynman model is correct to all orders in alpha, and the Hellwarth
+direct contour-integration of the general Feynman mobility statement is
+suitable for high temperature.
+
+It was necessary to return to these (rather old!) papers and resolve the
+models, as hybrid halide perovskites are soft materials with low energy
+phonons. Therefore the effective temperature in terms of a reduced
[…]
