+++ docs/src/examples.md
@@ -1,0 +1,204 @@
+Perhaps the easiest way to understand the code is to see how it can be used for
+science. 
+As an example system, we are going to look at some of the basic polaron
+properties of methylammonium lead-iodide perovskite.
+
+## Loading the Module
+
+However you run Julia (whether at the REPL; within a Jupyter notebook; or in
+a standalone script/program), you will first need to load the module. 
+As the Module is not yet registered, you will first need to tell Julia where
+to find the codes.
+
+```julia
+push!(LOAD_PATH,"../src/") # load module from local directory
+using PolaronMobility 
+```
+
+## α/alpha parameter 
+
+The Frohlich electron-phonon coupling parameter can be characterised by
+a dimensionless coupling, alpha (`α`). 
+This gives the long-range ('non analytic') contribution from electrodynamic
+coupling into infrared active phonon modes.
+
+As an example, let us calculate α for CdTe, and compare it to literature
+values.
+
+The call signature is (ϵ-optical, ϵ-static, phonon-frequency (Hz),
+effective-mass (in mass-of-electron units)).
+
+```
+α=feynmanalpha(7.1, 10.4, 5.08E12, 0.095)
+println("CdTe  α=",α," Stone 0.39 / Devreese 0.29")
+@test α ≈ 0.3 atol=0.1
+```
+
+We get a value of `0.351`. 
+
+## Single temperature phonon properties
+
+Let us calculate the room-temperature (300 K) character of the electron-polaron
+in methylammonium lead iodide perovskite (MAPI). 
+
+The call signature is (Temperature range, ϵ-optical, ϵ-static, phonon-frequency
+(Hz), effective-mass (in mass-of-electron units)). 
+For electrons in MAPI, these are ϵ=4.5/24.1, f=2.25 THz, me=0.12 electron
+masses. 
+
+
+```
+MAPIe=polaronmobility(300, 4.5, 24.1, 2.25E12, 0.12)
+```
+
+This will think for a bit (as Julia just-in-time compiles the required functions), and then spits out a considerable amount of information to `STDOUT`. 
+
+```
+Polaron mobility for system ε_Inf=4.5, ε_S=24.1, freq=2.25e12, 
+                 effectivemass=0.12; with Trange 300 ...
+Polaron mobility input parameters: ε_Inf=4.500000 ε_S=24.100000 freq=2.25e+12 α=2.393991 
+Derived params in SI: ω =1.41372e+13 mb=1.09313e-31 
+T: 300.000000 β: 2.41e+20 βred: 0.36 ħω  = 9.31 meV		Converged? : true
+ VariationalParams v= 19.86  w= 16.96	||	 M=0.371407 k=106.835753	
+ Polaron frequency (SI) v=  4.5e+13 Hz 	 w=  3.8e+13 Hz	
+ Feynman1955(46,47): meSmallAlpha(α)= 0.542 meLargeAlpha(α)= 0.067
+ Feynman1962: Approximate ~ Large alpha limit, v/w = 1.17  =~approx~= alpha^2 = 5.73 
+ POLARON SIZE (rf), following Schultz1959. (s.d. of Gaussian polaron ψ )
+	 Schultz1959(2.4): rf= 0.528075 (int units) = 2.68001e-09 m [SI]
+	 Schultz1959(2.5a) with 0.44: Feynman α→0 expansion: rfa= 1.68761 (int units) = 2.96691e-09 m [SI]
+	 Schultz1959(2.5a) with 4/9 re-derivation: Feynman α→0 expansion: rfa= 1.67915 (int units) = 2.95204e-09 m [SI]
[…]

[…]
