### 2018-02-20 cd09740
JOSS motivated docmentation improvements, part II - README integration.
