+++ docs/src/index.md
@@ -1,8 +1,19 @@
 These codes calculate the temperature-dependent polaron mobility for
-a material.
-We have parameters for various metal-halide Perovskite.
+a material. 
+We use the Feynman path-integral variational approach. 
+We have parameters for various metal-halide Perovskites; as well as other
+systems. 
+These codes implement methods described across a wide range of now quite old
+literature. 
+The methods have been tested against literature values, and the calculated
+units more well understood. 
+They enable relative 'turn key' calculation of polaron parameters (most
+particularly the finite temperature charge carrier mobility) for an
+arbitrary material system, based on parameters that are standard to calculate
+with modern ab-initio electronic structure methods. 
 
-It is perhaps easiest to read and understand this documentation alongside the paper:
+This documentation is intended to be read alongside the paper which was the
+first application (and motivation) for these codes: 
 [Frost2017PRB](https://doi.org/10.1103/PhysRevB.96.195202)
 ([ArXiv:1704.05404](https://arxiv.org/abs/1704.05404) ).
 
@@ -28,4 +39,42 @@ The codes are designed to produce a set of temperature-dependent mobilities and
 other data, for direct incorporation into a scientific publication.
 
 May your phonons drag in a manner truly sublime.
+
+## Installation
+
+These codes require Julia >0.6 . They are structured as a full Julia package,
+but are not yet registered with the central METADATA package repository. 
+
+To install, type the following at the Julia REPL:
+
+```
+julia> Pkg.clone("git://github.com/Jarvist/PolaronMobility.jl.git")
+```
+
+## Community guidelines
+
+Contributions to the code (extending that which is calculated), or additional
+physical systems / examples, are very welcome. 
+
+If you have questions about the software, scientific questions, or find errors,
+please create a [GitHub issue](https://github.com/jarvist/PolaronMobility.jl/issues). 
+
+If you find this package (or snippets, such as the entered and tested
+free-energy expressions) useful for your work, please cite the paper 
+[Frost2017PRB](https://doi.org/10.1103/PhysRevB.96.195202). 
+
+```
+@article{Frost2017,
+  doi = {10.1103/physrevb.96.195202},
+  url = {https://doi.org/10.1103/physrevb.96.195202},
+  year  = {2017},
+  month = {nov},
+  publisher = {American Physical Society ({APS})},
+  volume = {96},
+  number = {19},
+  author = {Jarvist Moore Frost},
+  title = {Calculating polaron mobility in halide perovskites},
+  journal = {Physical Review B}
+}
+```
 
