### 2018-02-20 f482ba8
Edits / corrections. Part of the JOSS documentation push.
