### 2018-02-22 jarvist/PolaronMobility.jl#1 @lheagy
Thanks @jarvist, @yxqd: would you also be willing to take a look?
