### 2018-02-26 jarvist/PolaronMobility.jl#1 @yxqd
Thanks @jarvist. The doc is much improved. For the examples, could you please add reference(s) for the parameters used in your calculation? For example, in your first example for CdTe, it says:
```
α=feynmanalpha(7.1, 10.4, 5.08E12, 0.095)
```
Please add a reference or two from which these parameters were obtained, maybe a volume of Landolt-Börnstein.
