+++ src/PolaronMobility.jl
@@ -1,6 +1,6 @@
 module PolaronMobility
 
-include("FeynmanKadanoffOsakaHellwarth.jl")
+include("FeynmanKadanoffOsakaHellwarth.jl") # Main polaronmobility function 
 include("Susceptibility.jl") # ImX calculation
 include("OedipusRex.jl") # Optical Absorption
 
