+++ src/FeynmanKadanoffOsakaHellwarth.jl
@@ -1,8 +1,8 @@
 # FeynmanKadanoffOsakaHellwarth.jl
 # Codes by Jarvist Moore Frost, 2017
 # Calculate Polaron Mobility - by a Osaka/Hellwarth variational solution to the Feynman model
-# If you run this, it should construct the model, solve for varying temperature, then produce plots as .pngs in the local directory.
-# These codes were developed with Julia 0.5.0, and requires the Optim and Plots packages.
+
+# These codes were developed with Julia 0.5.0 - Julia 0.6.2, and require the Optim and Plots packages.
 
 export Polaron # Type to hold the data
 export feynmanalpha, polaronmobility, savepolaron, plotpolaron
@@ -13,10 +13,6 @@ export ImX
 # one-dimensional numerical integration in Julia using adaptive Gauss-Kronrod quadrature
 import QuadGK.quadgk
 
-# Plot figures with Plots, which defaults to Pyplot backend
-#using Plots
-#pyplot()
-#default(size=(800,600)) # For the .png file output
 # Using the powerful Julia Optim package to optimise the variational parameters
 using Optim
 
@@ -27,14 +23,16 @@ const me=MassElectron = 9.10938188e-31;                          # kg
 const Boltzmann = const kB =  1.3806504e-23;                  # kg m2 / K s2 
 const ε_0 = 8.854E-12 #Units: C2N−1m−2, permittivity of free space
 
-# via Feynman 1955
-#   http://dx.doi.org/10.1103/PhysRev.97.660
 """
 feynmanalpha(ε_Inf,ε_S,freq,m_eff)
 
     Calculates the Frohlich alpha parameter, for a given dielectric constant,
     frequency (f) of phonon in Hertz, and effective mass (in units of the
     bare electron mass).
+
+    See Feynman 1955:
+    http://dx.doi.org/10.1103/PhysRev.97.660
+
 """
 function feynmanalpha(ε_Inf,ε_S,freq,m_eff)
     ω=freq*2*pi #frequency to angular velocity
@@ -47,15 +45,13 @@ function feynmanalpha(ε_Inf,ε_S,freq,m_eff)
 end
 
 #####
-# Hellwarth multiple phonon branch reduction
-
-# Most simple scheme
-# Hellwarth (58), assuming further typo on LHS, actually should be W_e
 """
 	HellwarthBScheme(LO)
+	
+    Multiple phonon mode reduction to a single effective frequency. 
+	Hellwarth et al. 1999 PRB, 'B scheme'; the athermal method. 
+    Averaging procedure is constructed by considering the average effect of the action of multiple branches. 
 
-	Multiple phonon mode reduction to a single effective frequency.
-	Hellwarth et al. 1999 PRB, 'B scheme'.
 	Follows Eqn (58) in this paper, assuming typo on LHS, should actually be W_e.
 """
 function HellwarthBScheme(LO)
@@ -78,11 +74,12 @@ end
 """
 	HellwarthAScheme(LO,T=295)
 
-	Multiple phonon mode reduction to a single effective frequency. 
+    Multiple phonon mode reduction to a single effective frequency. 
 	Temperature dependent, defaults to T=295 K.
+    
+    Follows Hellwarth et al. 1999 PRB 'A' scheme, Eqn 50 RHS.
+	
 	UNTESTED AND UNCERTAIN CODE.
-
-	Follows Hellwarth et al. 1999 PRB, Eqn 50 RHS.
 """
 function HellwarthAScheme(LO,T=295)
     println("Hellwarth A scheme...T=$T K")
[…]
