### 2018-04-01 8877993
FeynmanAthermal (orig 1955 version) vw numeric solution:
Also an example to show comparison with perturbative stuff,
Examples/FeynmanAthermalAsymptoticComparison.jl

Test cross-checks against Schultz numeric solutions:
test/FeynmanAthermal.jl

All this work was previously done in a Jupyter notebook ahead of my APS
March talk; see here:
https://github.com/jarvist/ijulia-notebooks/blob/master/2018-02-Athermal%20Feynman%20polarons.ipynb
