+++ README.md
@@ -1,26 +1,47 @@
-# PolaronMobility-FeynmanKadanoffOsakaHellwarth
+# PolaronMobility.jl
 
 [![Build Status](https://travis-ci.org/jarvist/PolaronMobility.jl.svg?branch=master)](https://travis-ci.org/jarvist/PolaronMobility.jl)
 [![Coverage Status](https://coveralls.io/repos/jarvist/PolaronMobility.jl/badge.svg?branch=master&service=github)](https://coveralls.io/github/jarvist/PolaronMobility.jl?branch=master)
 [![codecov.io](http://codecov.io/github/jarvist/PolaronMobility.jl/coverage.svg?branch=master)](http://codecov.io/github/jarvist/PolaronMobility.jl?branch=master)
 
 These codes calculate the temperature-dependent polaron mobility for
-a material.
-We have parameters for various metal-halide Perovskite.
+a material. 
+This is based on the Feynman variational solution to the Polaron problem. 
+The electron-phonon coupling is treated as an effective α (alpha) Frohlich
+Hamiltonian parameter. 
+The band structure is treated with an effective mass theory. 
+The variational problem is solved numerically for finite-temperature free
+energies, rather than by the asymptotic solutions to the athermal energies, as
+often found in the textbooks.  
+The mobility is calculated numerically by integrating the polaron self-energy
+along the imaginary axis, as well as with the Kadanoff Boltzmann equation
+approximation, and the FHIP low-temperature asymptotic solution. 
+
+We provided parameters for various metal-halide Perovskites, and other
+interesting systems.
+
+The motivation for developing these codes was to enable polaron mobility
+calculations on arbitrary materials. 
+They also provide the only extant implementation of Feynman's variational
+method.  
+They offer a convenient basis for writing codes that build on these variational
+solutions. 
 
 More extensive documentation is available
 [![here](https://img.shields.io/badge/docs-latest-blue.svg)](https://jarvist.github.io/PolaronMobility.jl/),
-which is perhaps easiest to read and understand alongside the pre-print:
-[ArXiv:1704.05404](https://arxiv.org/abs/1704.05404) .
+which is perhaps easiest to read and understand alongside the first paper:
+[ArXiv:1704.05404](https://arxiv.org/abs/1704.05404)
+/ [Frost2017PRB](https://doi.org/10.1103/PhysRevB.96.195202).
 
-The required inputs are the dielectric constants (ϵ-static and ϵ-optic)
-, a characteristic phonon frequency (ω), and the bare-electron band
-effective-mass (me).  These values can be relatively easily calculated in the
+The model inputs are the dielectric constants (ϵ-static and ϵ-optic),
+a characteristic phonon frequency (ω), and the bare-electron band
+effective-mass (me).  
+These values can be relatively easily calculated in the
 ab-initio electronic structure package of your choosing, or measured directly. 
 
 From these four values, the code solves a temperature-dependent polaron model. 
-This is done by variationally optimising the temperature-dependent free-energies for the coupled
-electron-phonon system. 
+This is done by variationally optimising the temperature-dependent
+free-energies for the coupled electron-phonon system. 
 These optimised parameters describe the polaron with the infinite quantum field
 of lattice vibrations 'integrated through', and replaced with a phonon-drag
 term. 
@@ -35,7 +56,7 @@ other data, for direct incorporation into a scientific publication.
 
 May your phonons drag in a manner truly sublime.
 
-![MAPI Polaron mobility, plotted vs expt data](mobility-calculated-experimental.png)
+![MAPI Polaron mobility, plotted vs experimental data](mobility-calculated-experimental.png)
 
 ## Installation
 
[…]
