### 2018-04-01 1b65151
Documentation / README clean + unicode niceify.
