### 2018-04-01 ab3f3e7
Add to automated documentation, cleanup, reorder.
