### 2018-04-01 #566 @jarvist
Hello, sorry things have dragged on a bit - I was doing a fair bit of travelling & moved institution.

I think I've covered most of the issues highlighted with the documentation, and have written an extended 'Examples' component which I hope will help teach the physics. I also believe I now hit the standard JOSS requirements, including a section in the README on contributing.

I had a look at code coverage, but it's already at 100% of the code actually used to generate results! That which is not covered is a work in progress, which might be considerably refactored.

Best.
