### 2018-04-10 jarvist/PolaronMobility.jl#1 @yxqd
Thanks @jarvist. The documentation is much improved! Here are some small problems to fix

# Installation
Please replace "git://" with "https://". The former fails if there is firewall. It is safer to assume https works for most people.

# Examples
* `push!(LOAD_PATH,"../src/") # load module from local directory`
  Is it necessary if the installation step is already performed? Is this for "developers" rather than "users"?
* `@test` failed. `using Base.Test` fixed that. Please add a hint in the documentation
* using PlotPolaron errored with the following message:
   ```ERROR: ArgumentError: Module PlotPolaron not found in current path.
    Run `Pkg.add("PlotPolaron")` to install the PlotPolaron package.
    Stacktrace:
     [1] _require(::Symbol) at ./loading.jl:435
     [2] require(::Symbol) at ./loading.jl:405
    ```
* as suggested by @lheagy , could you add some pictures of typical plots expected for the **Temperature-dependent behaviour** example?
