### 2018-04-10 jarvist/PolaronMobility.jl#2 @yxqd
JOSS Review: README.md
@jarvist Please consider simplify the README.md. Some of the information there is now available at the documentation site https://jarvist.github.io/PolaronMobility.jl/ and looks duplicate. It would be nice that README.md is a succinct summary and defer more elaborate discussions to the documentation site. [This](https://github.com/ropensci/hydroscoper/blob/052a37e96255feef86c4ad1a38a6517db4d6e11e/README.md) might be a good example.

See also: #1, openjournals/joss-reviews#566
