### 2018-04-10 #566 @yxqd
I think the documentation is much improved now.  There are some small problems and comments were added to jarvist/PolaronMobility.jl#1 and  jarvist/PolaronMobility.jl#2
