### 2018-04-12 f6c8a9b
Rewrite and simplify README.
