+++ README.md
@@ -4,8 +4,9 @@
 [![Coverage Status](https://coveralls.io/repos/jarvist/PolaronMobility.jl/badge.svg?branch=master&service=github)](https://coveralls.io/github/jarvist/PolaronMobility.jl?branch=master)
 [![codecov.io](http://codecov.io/github/jarvist/PolaronMobility.jl/coverage.svg?branch=master)](http://codecov.io/github/jarvist/PolaronMobility.jl?branch=master)
 
-These codes calculate the temperature-dependent polaron mobility for
-a material. 
+`PolaronMobility.jl` is a Julia package which calculates the
+temperature-dependent polaron mobility for a material. 
+
 This is based on the Feynman variational solution to the Polaron problem. 
 The electron-phonon coupling is treated as an effective α (alpha) Frohlich
 Hamiltonian parameter. 
@@ -33,58 +34,44 @@ which is perhaps easiest to read and understand alongside the first paper:
 [ArXiv:1704.05404](https://arxiv.org/abs/1704.05404)
 / [Frost2017PRB](https://doi.org/10.1103/PhysRevB.96.195202).
 
-The model inputs are the dielectric constants (ϵ-static and ϵ-optic),
-a characteristic phonon frequency (ω), and the bare-electron band
-effective-mass (me).  
-These values can be relatively easily calculated in the
-ab-initio electronic structure package of your choosing, or measured directly. 
-
-From these four values, the code solves a temperature-dependent polaron model. 
-This is done by variationally optimising the temperature-dependent
-free-energies for the coupled electron-phonon system. 
-These optimised parameters describe the polaron with the infinite quantum field
-of lattice vibrations 'integrated through', and replaced with a phonon-drag
-term. 
-From this the polaron features such as effective-mass, size of the
-wavefunction, frequency of energy oscillation etc. can be calculated. 
-
-This polaron state can then be used as an input to further models for polaron
-mobility. 
-
-The codes are designed to produce a set of temperature-dependent mobilities and
-other data, for direct incorporation into a scientific publication. 
-
-May your phonons drag in a manner truly sublime.
-
-![MAPI Polaron mobility, plotted vs experimental data](mobility-calculated-experimental.png)
 
 ## Installation
 
-These codes require Julia >0.6 . They are structured as a full Julia package, but are not yet registered with the central METADATA package repository. 
+These codes require Julia >0.6 . They are structured as a full Julia package,
+but are not yet registered with the central METADATA package repository. 
 
 To install, type the following at the Julia REPL:
 
 ```
-julia> Pkg.clone("git://github.com/Jarvist/PolaronMobility.jl.git")
+julia> Pkg.clone("https://github.com/Jarvist/PolaronMobility.jl.git")
 ```
+
+## Using
+
+As an example:
+
+```
+using PolaronMobility
+MAPIe=polaronmobility(300, 4.5, 24.1, 2.25E12, 0.12)
+```
+
+Will calculate the polaron mobility for methyl-ammonium lead halide perovskite
+(f=2.25 THz; ϵoptical=4.5; ϵstatic=24.1; effective-mass=0.12 electron-masses)  
+at 300 K. 
+
+Further details in the
+[documentation](https://jarvist.github.io/PolaronMobility.jl/).
 
 ## Research outputs
 
-As an example, here are some indicative figures from an early version of the codes. 
-These are specified for transport in hybrid lead iodide perovskite. 
+The central output of this model are temperature-dependent polaron mobilities: 
 
-Polaron mobilities, three different ways
-![Polaron mobilities, three different ways](mobility-calculated.png)
+![MAPI Polaron mobility, plotted vs experimental data](mobility-calculated-experimental.png)
 
-Effective mass of phonon cloud
-![Effective mass of phonon cloud](mass.png)
-
[…]
