### 2018-04-12 7d09de6
Typical output.
