### 2018-04-26 0a7d1e9
Correct all instances feynmanalpha -> frohlichalpha.
