### 2018-04-26 19f428a
Moar oxides; from the Materials Project.
