### 2018-04-29 #566 @lheagy
👋 Hi @jarvist: just wanted to check in and see how things are going addressing jarvist/PolaronMobility.jl#1 and jarvist/PolaronMobility.jl#2
