### 2018-05-01 9fb72be
Simplify output, and move less interesting data -> verbose only.
