+++ src/FeynmanKadanoffOsakaHellwarth.jl
@@ -149,7 +149,7 @@ AF(v,w,α)=π^(-0.5) * α*v * quadgk(τ->fF(τ,v,w),0,Inf)[1]
 # (33) in Feynman I
 F(v,w,α)=(3/(4*v))*(v-w)^2-AF(v,w,α)
 
-# Let's wrap this in a simple function
+# Let's wrap the Feynman athermal variation approximation in a simple function
 """
     feynmanvw(α)
 
@@ -276,7 +276,7 @@ function polaronmobility(Trange, ε_Inf, ε_S, freq, effectivemass; verbose::Boo
         v=minimum[1] # unpack these from minimised results
         w=minimum[2]
             
-        @printf("\n VariationalParams v= %.2f  w= %.2f",v,w)
+        @printf("\n Polaraon Parameters:  v= %.4f  w= %.4f",v,w)
 
         # From 1962 Feynman, definition of v and w in terms of the coupled Mass and spring-constant
         # See Page 1007, just after equation (18)
@@ -287,20 +287,25 @@ function polaronmobility(Trange, ε_Inf, ε_S, freq, effectivemass; verbose::Boo
         append!(p.k,k)
         append!(p.M,M)
 
-        @printf("\t||\t M=%f k=%f\t",M,k)
+        @printf("  ||   M=%f  k=%f\t",M,k)
        
-        @printf("\n Polaron frequency (SI) v=  %.2g Hz \t w=  %.2g Hz\t",v*ω /(2*pi),w*ω /(2*pi))
+        @printf("\n Polaron frequency (SI) v= %.2g Hz  w= %.2g Hz",
+                v*ω /(2*pi),w*ω /(2*pi))
 
 		# (46) in Feynman1955
 		meSmallAlpha(α )=α /6 + 0.025*α ^2
 		# (47) In Feynman1955
 		meLargeAlpha(α )=16*α ^4 / (81*π ^4)
 		#meLargeAlpha(α )=202*(α /10)^4
-		@printf("\n Feynman1955(46,47): meSmallAlpha(α)= %.3f meLargeAlpha(α)= %.3f",meSmallAlpha(α),meLargeAlpha(α))
-        @printf("\n Feynman1962: Approximate ~ Large alpha limit, v/w = %.2f  =~approx~= alpha^2 = %.2f ",v/w,α ^2)
+        if (verbose) # asymptotic solutions - not that interesting when you have the actual ones!
+		    @printf("\n Feynman1955(46,47): meSmallAlpha(α)= %.3f meLargeAlpha(α)= %.3f",
+                    meSmallAlpha(α),meLargeAlpha(α))
+            @printf("\n Feynman1962: Approximate ~ Large alpha limit, v/w = %.2f  =~approx~= alpha^2 = %.2f ",
+                    v/w,α ^2)
+        end
 
         # POLARON SIZE
-        @printf("\n POLARON SIZE (rf), following Schultz1959. (s.d. of Gaussian polaron ψ )")
+        @printf("\n Polaron size (rf), following Schultz1959. (s.d. of Gaussian polaron ψ )")
         # Schultz1959 - rather nicely he actually specifies everything down into units!
         # just before (2.4) in Schultz1959
         mu=((v^2-w^2)/v^2)
@@ -311,6 +316,7 @@ function polaronmobility(Trange, ε_Inf, ε_S, freq, effectivemass; verbose::Boo
         @printf("\n\t Schultz1959(2.4): rf= %g (int units) = %g m [SI]",rf,rfsi )
         append!(p.rfsi,rfsi)
 
+        if (verbose)
         scale=sqrt(2*mb*ω) # Note we're using mb; 
[…]
