### 2018-05-01 d3b46b9
Tidy up; doing more serious study in separate repository.
