### 2018-05-02 jarvist/PolaronMobility.jl#2 @jarvist
Hopefully https://github.com/jarvist/PolaronMobility.jl/commit/f6c8a9b616abf4240bc9bdae84007c4b7639d23f#diff-04c6e90faac2675aa89e2176d2eec7d8 and https://github.com/jarvist/PolaronMobility.jl/commit/7d09de64704af5b0652b61a6adf6b0d0ecd83a75#diff-04c6e90faac2675aa89e2176d2eec7d8 achieve this.
If that's sufficient, could you close the issue?
