### 2018-05-02 #566 @jarvist
https://github.com/jarvist/PolaronMobility.jl/issues/2 is hopefully dealt with.
https://github.com/jarvist/PolaronMobility.jl/issues/1 requires a bit more work still - just need to make the time to finish a gentle refactor + ensure the documentation is current. Thanks for the reminder!
