### 2018-05-14 e4ac60d
Improvements to documentation; re-sync to latest interface changes.
