+++ docs/src/examples.md
@@ -5,14 +5,18 @@ properties of methylammonium lead-iodide perovskite.
 
 ## Loading the Module
 
-However you run Julia (whether at the REPL; within a Jupyter notebook; or in
-a standalone script/program), you will first need to load the module. 
-As the Module is not yet registered, you will first need to tell Julia where
-to find the codes.
+First just run an import of the module, to see if Julia can find it.
+
+```julia
+using PolaronMobility 
+```
+
+If you are running the module from outside the Julia package directory (i.e.
+you have cloned the repository elsewhere to more easily work on the codes), you
+can supplement the `LOAD_PATH`.
 
 ```julia
 push!(LOAD_PATH,"../src/") # load module from local directory
-using PolaronMobility 
 ```
 
 ## α/alpha parameter 
@@ -22,11 +26,21 @@ a dimensionless coupling, alpha (`α`).
 This gives the long-range ('non analytic') contribution from electrodynamic
 coupling into infrared active phonon modes.
 
-As an example, let us calculate α for CdTe, and compare it to literature
+```math
+\alpha = 
+\frac{1}{2} \;
+\frac{1}{4\pi\epsilon_0} \:
+\left( \frac{1}{\epsilon_{optical}} - \frac{1}{\epsilon_{static}} \right ) \;
+\frac{e^2}{\hbar \omega} \;
+\sqrt{\frac{2m_e\omega}{\hbar}}
+```
+
+This is provided as a convenience function (with correct units!). 
+Let us demonstrate by calculating α for CdTe, and compare it to literature
 values.
 
-The call signature is (ϵ-optical, ϵ-static, phonon-frequency (Hz),
-effective-mass (in mass-of-electron units)).
+The call signature is: ϵ-optical, ϵ-static, phonon-frequency (Hz),
+effective-mass (in mass-of-electron units).
 
 ```
 α=frohlichalpha(7.1, 10.4, 5.08E12, 0.095)
@@ -40,10 +54,12 @@ We get a value of `0.351`.
 
 Let us calculate the room-temperature (300 K) character of the electron-polaron
 in methylammonium lead iodide perovskite (MAPI). 
-The parameters we use are as in Frost2017PRB.
+The parameters we use are as in `Frost2017PRB`.
 
-The call signature is (Temperature range, ϵ-optical, ϵ-static, phonon-frequency
-(Hz), effective-mass (in mass-of-electron units)). 
+The call signature to `polaronmobility` is: 
+Temperature range, ϵ-optical, ϵ-static, phonon-frequency (Hz), effective-mass
+(in mass-of-electron units). 
+
 For electrons in MAPI, these are ϵ=4.5/24.1, f=2.25 THz, me=0.12 electron
 masses. 
 
@@ -62,53 +78,44 @@ Polaron mobility for system ε_Inf=4.5, ε_S=24.1, freq=2.25e12,
 Polaron mobility input parameters: ε_Inf=4.500000 ε_S=24.100000 freq=2.25e+12 α=2.393991 
 Derived params in SI: ω =1.41372e+13 mb=1.09313e-31 
 T: 300.000000 β: 2.41e+20 βred: 0.36 ħω  = 9.31 meV		Converged? : true
- VariationalParams v= 19.86  w= 16.96	||	 M=0.371407 k=106.835753	
- Polaron frequency (SI) v=  4.5e+13 Hz 	 w=  3.8e+13 Hz	
- Feynman1955(46,47): meSmallAlpha(α)= 0.542 meLargeAlpha(α)= 0.067
- Feynman1962: Approximate ~ Large alpha limit, v/w = 1.17  =~approx~= alpha^2 = 5.73 
- POLARON SIZE (rf), following Schultz1959. (s.d. of Gaussian polaron ψ )
[…]
