### 2018-05-14 e95d5bb
Update docs; example figure; more info on PlotPolaron
