### 2018-05-14 jarvist/PolaronMobility.jl#1 @jarvist
Hello everyone - hopefully with this last commit (https://github.com/jarvist/PolaronMobility.jl/commit/e4ac60d35d0b0e67efb748c5789b0245bfa897da), all of the issues above should have been solved!

Longer term the package could definitely do with some restructing, breaking down of the fairly monolithic functions, and a complete writing of the `PlotPolaron` pseudo-package in terms of `Plots.jl` recipes. But it will be easier to do these more organically while also extending + improving the package.
