### 2018-05-14 #566 @jarvist
Hi everybody,
Hopefully https://github.com/jarvist/PolaronMobility.jl/issues/1 should be resolved with the latest round of commits culminating in https://github.com/jarvist/PolaronMobility.jl/commit/e4ac60d35d0b0e67efb748c5789b0245bfa897da.

Longer-term, there's definitely some refactoring + more documentation that would be good to write. However, I think that the package is usable in it's current state, and it should be possible to organically improve on it from this point onwards.
