### 2018-05-18 jarvist/PolaronMobility.jl#1 @yxqd
@jarvist :+1: with the improvements of documentation on usage and the plots.

The instructions on plotting using PlotPolaron still does not work for me, but the alternative (gnuplot) works.

I feel the part on using PlotPolaron can be made smoother for users. But I am not an expert in Julia, and don't know enough to comment. @lheagy please decide if you want to ask an Julia expert to review or leave it as is. In the latter case, I would suggest @jarvist to make the gnuplot instruction more complete, but it is not required for closing this issue.
