### 2018-05-18 #566 @yxqd
@lheagy thanks for the reminder. Just left my comments at  jarvist/PolaronMobility.jl#1. Please review
