### 2018-05-22 jarvist/PolaronMobility.jl#1 @jarvist
OK. That all makes sense. I'll try and fix the plotting one-way or another.
What I really need to do is write some Plots.jl recipes, but this I something I do not yet understand how to do.
