### 2018-05-23 jarvist/PolaronMobility.jl#1 @wkearn
Hello all! @Kevin-Mattheus-Moerman pointed me to your discussion here, and I'm happy to offer any advice re: plotting that I can.

This is what I had to do to get your plots to run:

```julia
julia> using PolaronMobility, Plots
julia> gr()
julia> include(Pkg.dir("PolaronMobility","src","PlotPolaron.jl"))
julia> using PlotPolaron
julia> ... # Do your polaronmobility calculations
julia> plotpolaron(...)
```
This will work if you have installed `PolaronMobility.jl` using `Pkg.clone` and you have `Plots.jl` installed. The `using Plots` and `gr()` calls are necessary on my machine to get around a world age error, but I'm not sure if that is something wrong here or in Plots. I'll try to track down that bug.

Also when I try to plot the example

```
julia> MAPIe=polaronmobility(10:10:1000, 4.5, 24.1, 2.25E12, 0.12)
julia> plotpolaron("MAPI-electron",MAPIe)
```

I get an error in the radius vs temperature plot because `MAPIe.rfsmallalpha` doesn't have anything in it. The plots before that one work fine, I think, but I don't have enough domain knowledge here to check.

@jarvist if you want some help converting `plotpolaron` into `Plots.jl` recipes, let me know (tag me in a new issue or something). At a first glance you aren't doing anything too complicated with your plots, so this should be fairly straightforward once you get over the admittedly steep learning curve of recipes.
