### 2018-06-05 fe95ac7
Split thermal v,w off into own part of code. Should allow T=0 mobility.
TODO: Make athermal v,w an option for T-dep mobility.
Check that T=0 values make sense for the T->0 trend.
Tidy up and reduce code - should probably generalise the feynmanvw code
more.
