+++ src/FeynmanKadanoffOsakaHellwarth.jl
@@ -17,10 +17,10 @@ import QuadGK.quadgk
 using Optim
 
 # Physical constants
-const hbar = const ħ = 1.05457162825e-34;          # kg m2 / s 
-const eV = const q = const ElectronVolt = 1.602176487e-19;                         # kg m2 / s2 
+const hbar = const ħ = 1.05457162825e-34;          # kg m2 / s
+const eV = const q = const ElectronVolt = 1.602176487e-19;                         # kg m2 / s2
 const me=MassElectron = 9.10938188e-31;                          # kg
-const Boltzmann = const kB =  1.3806504e-23;                  # kg m2 / K s2 
+const Boltzmann = const kB =  1.3806504e-23;                  # kg m2 / K s2
 const ε_0 = 8.854E-12 #Units: C2N−1m−2, permittivity of free space
 
 """
@@ -36,12 +36,12 @@ frohlichalpha(ε_Inf,ε_S,freq,m_eff)
 """
 function frohlichalpha(ε_optic,ε_static,freq,m_eff)
     ω=freq*2*pi #frequency to angular velocity
-    # Note: we need to add a 4*pi factor to the permitivity of freespace. 
+    # Note: we need to add a 4*pi factor to the permitivity of freespace.
     # This gives numeric agreement with literature values.  This is required as
     # the contemporary 1950s and 1960s literature implicitly used atomic units,
     # where the electric constant ^-1 has this factor baked in, k_e=1/(4πϵ_0).
     α= 0.5/(4*π*ε_0) *              # Units: m/F
-       (1/ε_optic - 1/ε_static) *   # Units: none 
+       (1/ε_optic - 1/ε_static) *   # Units: none
        (q^2/(hbar*ω)) *             # Units: F
        sqrt(2*me*m_eff*ω/ħ)         # Units: 1/m
     return (α)
@@ -50,10 +50,10 @@ end
 #####
 """
 	HellwarthBScheme(LO)
-	
-    Multiple phonon mode reduction to a single effective frequency. 
-	Hellwarth et al. 1999 PRB, 'B scheme'; the athermal method. 
-    Averaging procedure is constructed by considering the average effect of the action of multiple branches. 
+
+    Multiple phonon mode reduction to a single effective frequency.
+	Hellwarth et al. 1999 PRB, 'B scheme'; the athermal method.
+    Averaging procedure is constructed by considering the average effect of the action of multiple branches.
 
 	Follows Eqn (58) in this paper, assuming typo on LHS, should actually be W_e.
 """
@@ -77,11 +77,11 @@ end
 """
 	HellwarthAScheme(LO,T=295)
 
-    Multiple phonon mode reduction to a single effective frequency. 
+    Multiple phonon mode reduction to a single effective frequency.
 	Temperature dependent, defaults to T=295 K.
-    
+
     Follows Hellwarth et al. 1999 PRB 'A' scheme, Eqn 50 RHS.
-	
+
 	UNTESTED AND UNCERTAIN CODE.
 """
 function HellwarthAScheme(LO,T=295)
@@ -132,14 +132,14 @@ B(v,w,β,α) = α*v/(sqrt(π)*(exp(β)-1)) * quadgk(x->f(x,v,w,β),0,β/2)[1]
 # 62e
 C(v,w,β)=3/4*(v^2-w^2)/v * (coth(v*β/2)-2/(v*β))
 # 62a
-F(v,w,β,α)=-(A(v,w,β)+B(v,w,β,α)+C(v,w,β)) 
+F(v,w,β,α)=-(A(v,w,β)+B(v,w,β,α)+C(v,w,β))
 
 # Can now evaluate, e.g.
 # F(v,w,β,α)=F(7.2,6.5,1.0,1.0)
[…]
