### 2018-06-29 7691566
Hot fix for Optim 0.15- compatability; change of call signature.
