### 2018-06-29 daa4e57
Documentation updates.
