+++ docs/mkdocs.yml
@@ -6,7 +6,7 @@ pages:
     - Examples: examples.md  
     - Functions provided: functions.md  
 
-theme: readthedocs
+theme: windmill 
 
 docs_dir: build # from Julia's Documentation.jl
 
