### 2018-06-29 4d492b8
Comment out obsolete rfsmallalpha approx; no longer calculated.
