### 2018-06-29 jarvist/PolaronMobility.jl#1 @jarvist
Thanks for that @wkearn . I've added your instructions to the documentation, and fixed (by commenting out) the obsolete attempt to plot rfsmallalpha.

Hopefully we should be good to go I think!
