### 2018-06-29 #566 @jarvist
Hi @lheagy  - thanks for the reminder! I've updated the codes to work with the latest upstream packages, and integrated wkearn's suggestion for how to get PlotPolaron working on an ad-hoc basis at least.
I think getting the inbuilt plotting working for users was the last sticking point? If you are happy to close https://github.com/jarvist/PolaronMobility.jl/issues/1 with https://github.com/jarvist/PolaronMobility.jl/commit/daa4e5789bbb6a6be102cc57b7f9aa9304b3025d#diff-3fb8eca8bec7587dc56e13745886617cL214 , I am as well!
