### 2018-06-29 07460ab
Documenting the Jupyter notebook for plotting.
