### 2018-07-01 #566 @lheagy
Thanks @jarvist.
👋 @yxqd: Thanks for all of your work on this review. Would you be willing to take another look at jarvist/PolaronMobility.jl#1, and if you see that it has been addressed, feel free to close the issue. Thanks!
