### 2018-07-04 jarvist/PolaronMobility.jl#1 @yxqd
Using Juliabox is a nice idea. But I encountered this problem:

![image](https://user-images.githubusercontent.com/1796155/42293418-0bdfb174-7fa8-11e8-9ffd-a48895ba417a.png)

And I cannot test the plotting in juliabox either since the calculations failed at cell 4
