### 2018-07-04 jarvist/PolaronMobility.jl#1 @jarvist
How frustrating! (...After a bit of investigation...) This is because JuliaBox is still on `Optim 0.14` , and it does not pay regard to the REQUIRE file when you clone a package.
I updated the function signature last Friday with https://github.com/jarvist/PolaronMobility.jl/commit/769156613c77abbc9a428c345576350c1bfaf56a , so that the package works from the latest `Julia 0.6.x` with a `Pkg.update()` (which pulls in the new `Optim 0.15` interface).

As a work around, you can add 'Optim' to the 'Yours' list in the 'Packages' menu on the root of JuliaBox. It takes a while to build the package; and then a while to compile after you restart your kernel. This appears to pull in a fresh stable copy.
