### 2018-07-09 jarvist/PolaronMobility.jl#1 @yxqd
Thanks. I tried the command line interface locally, and it does work now after `Pkg.update()`. Maybe you can add these little problems somewhere in the documentation (maybe a "Known problems" section), or as github issues, so that you can easily refer users to those hints if this comes up again.

The notebook at JuliaBox mostly works now after adding Optim to "registered" packages of JuliaBox. Maybe you want to add this somewhere in your documentation too.

There is still one last error with the JuliaBox notebook:

![image](https://user-images.githubusercontent.com/1796155/42464068-f18f4042-8375-11e8-8a28-df499d97a3a0.png)

Please check.
