### 2018-07-13 2774404
Restructure package into sensible source files.
