+++ src/FeynmanTheory.jl
@@ -25,35 +25,8 @@ function frohlichalpha(ε_optic,ε_static,freq,m_eff)
 end
 
 #####
+# Athermal (Feynman 1955) model
 # Set up equations for the polaron free energy, which we will variationally improve upon
-
-# Hellwarth et al. 1999 PRB - Part IV; T-dep of the Feynman variation parameter
-
-# Originally a Friday afternoon of hacking to try and implement the T-dep electron-phonon coupling from the above PRB
-# Which was unusually successful! And more or less reproduced Table III
-
-# Define Osaka's free-energies (Hellwarth1999 version) as Julia functions
-# Equation numbers follow above Hellwarth et al. 1999 PRB
-# 62b
-A(v,w,β)=3/β*( log(v/w) - 1/2*log(2*π*β) - log(sinh(v*β/2)/sinh(w*β/2)))
-# 62d
-Y(x,v,β)=1/(1-exp(-v*β))*(1+exp(-v*β)-exp(-v*x)-exp(v*(x-β)))
-# 62c integrand
-f(x,v,w,β)=(exp(β-x)+exp(x))/(w^2*x*(1-x/β)+Y(x,v,β)*(v^2-w^2)/v)^(1/2)
-# 62c
-B(v,w,β,α) = α*v/(sqrt(π)*(exp(β)-1)) * quadgk(x->f(x,v,w,β),0,β/2)[1]
-# 62e
-C(v,w,β)=3/4*(v^2-w^2)/v * (coth(v*β/2)-2/(v*β))
-# 62a
-F(v,w,β,α)=-(A(v,w,β)+B(v,w,β,α)+C(v,w,β))
-
-# Can now evaluate, e.g.
-# F(v,w,β,α)=F(7.2,6.5,1.0,1.0)
-# BUT - this is just the objective function! (The temperature-dependent free-energy.) Not the optimised parameters.
-# Also there's a scary numeric integration (quadgk) buried within...
-
-# In Julia we have 'Multiple dispatch', so let's just construct the Feynman (athermal)
-# energy as the same signature, but without the thermodynamic beta
 
 # Integrand of (31) in Feynman I (Feynman 1955, Physical Review, "Slow electrons...")
 fF(τ,v,w)=(w^2 * τ + (v^2-w^2)/v*(1-exp(-v*τ)))^-0.5 * exp(-τ)
@@ -88,6 +61,37 @@ function feynmanvw(α; v=7.0, w=6.0) # v,w defaults
 
     return v,w
 end
+
+
+
+# Hellwarth et al. 1999 PRB - Part IV; T-dep of the Feynman variation parameter
+
+# Originally a Friday afternoon of hacking to try and implement the T-dep electron-phonon coupling from the above PRB
+# Which was unusually successful! And more or less reproduced Table III
+
+# In Julia we have 'Multiple dispatch', so let's just construct the free
+# energies (temperature-dependent) with the same name as before, but withthe
+# thermodynamic beta where required
+
+# Define Osaka's free-energies (Hellwarth1999 version) as Julia functions
+# Equation numbers follow above Hellwarth et al. 1999 PRB
+# 62b
+A(v,w,β)=3/β*( log(v/w) - 1/2*log(2*π*β) - log(sinh(v*β/2)/sinh(w*β/2)))
+# 62d
+Y(x,v,β)=1/(1-exp(-v*β))*(1+exp(-v*β)-exp(-v*x)-exp(v*(x-β)))
+# 62c integrand
[…]
