### 2018-07-14 d666222
Chipping away at Finite temperature HellwarthA scheme. Confused!
