### 2018-08-03 e3827f2
Add notes on Optim.jl versions; Juliabox still defaulting to old one.
