### 2018-08-03 jarvist/PolaronMobility.jl#1 @jarvist
@yxqd OK Jiao, I've pushed a new version of the ipython notebook without the 'integer division' error (the approximation is no longer calculated as standard).

I've also introduced some text to `README.md` suggesting that users install the `Optim` package themselves on Juliabox. Hopefully within a few weeks the standard images on Juliabox will update.

I think this should be sufficient to cover the final checkbox above - "Example usage".
