### 2018-08-03 #566 @jarvist
I've just commited changes which I hope will enable issue https://github.com/jarvist/PolaronMobility.jl/issues/1 to be closed.

Are there any further blocking corrections that are wished for 'Functionality' or 'Documentation' in the tick-boxes above?
