### 2018-08-04 #566 @yxqd
@lheagy my review is done. Please also close https://github.com/jarvist/PolaronMobility.jl/issues/1 if you see fit.
