### 2018-08-05 jarvist/PolaronMobility.jl#1 @lheagy
This looks great @jarvist!
