### 2018-08-05 #566 @lheagy
Many thanks @yxqd for taking the time on this review! Your input is much appreciated.

@jarvist: could you please archive the software on [zenodo](https://zenodo.org/) or similar and share the doi here?
