### 2018-08-05 #566 @jarvist
Dear Both,
Many thanks!

https://github.com/jarvist/PolaronMobility.jl/releases/tag/JOSS
DOI: 10.5281/zenodo.1336754
